requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
alembic
pydantic-settings
python-dotenv